Remove Server interface
