+++ src/test/java/com/cafiaso/server/network/protocol/packets/server/status/StatusResponsePacketTest.java
@@ -5,6 +5,10 @@ import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 
+import static com.cafiaso.server.Server.MINECRAFT_VERSION;
+import static com.cafiaso.server.Server.PROTOCOL_VERSION;
+import static com.cafiaso.server.configuration.ServerConfiguration.DEFAULT_DESCRIPTION;
+import static com.cafiaso.server.configuration.ServerConfiguration.DEFAULT_MAX_PLAYERS;
 import static com.cafiaso.server.network.protocol.packets.server.status.StatusResponsePacket.*;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -12,7 +16,8 @@ class StatusResponsePacketTest {
 
     @Test
     void serialize_OK() throws IOException {
-        String jsonResponse = "{\"players\":{\"max\":20,\"online\":10},\"description\":{\"text\":\"A Minecraft Server\"},\"version\":{\"protocol\":770,\"name\":\"1.21.5\"}}";
+        String jsonResponse = "{\"players\":{\"max\":%d,\"online\":10},\"description\":{\"text\":\"%s\"},\"version\":{\"protocol\":%d,\"name\":\"%s\"}}"
+                .formatted(DEFAULT_MAX_PLAYERS, DEFAULT_DESCRIPTION, PROTOCOL_VERSION, MINECRAFT_VERSION);
 
         FriendlyBuffer buffer = new FriendlyBuffer();
